requirements.txt
transformers==4.36.0
torch==2.1.0
faiss-cpu==1.7.4
openai==1.3.0
newspaper3k==0.2.8
lxml[html_clean]==4.9.3
lxml_html_clean==0.1.1
praw==7.7.1
requests==2.31.0
beautifulsoup4==4.12.2
streamlit==1.28.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
python-dotenv==1.0.0
sentence-transformers==2.2.2